[broker-67] Introduce AuthenticationSpringConfig
