+++ application/src/test/resources/auth/user-credentials-data.sql
@@ -1,2 +1,1 @@
 INSERT INTO user_credentials(username, password) VALUES ('user', X'01 bc 2a');
-
