+++ credentials-source-db/build.gradle
@@ -9,9 +9,9 @@ description = "Database-based Credentials Source Provider"
 dependencies {
   api projects.base
   api projects.authenticationBasic
-  implementation libs.r2dbc.spi
-  implementation libs.r2dbc.pool
-  implementation libs.r2dbc.postgresql
+  api libs.r2dbc.spi
+  api libs.r2dbc.pool
+  api libs.r2dbc.postgresql
 
   testImplementation projects.testSupport
   testFixturesApi projects.testSupport
